Update multi_port_balance_usdt.py
Fine tuned the trailing orders to perform better in most cases. Trailing orders are tricky, that's why I am splitting the buy/sell operations in two orders (when possible) and watch which one performs better. In any case the two trailing orders make a very good mini-DCA, an awesome average buy or sell price.
Have fun and use with care!
I'm not a financial advisor. Past performance is not guarantee of future gains. Use with care.